requirements.txt
streamlit>=1.26.0
numpy>=1.24.0
pillow>=9.5.0
deepface>=0.0.92
transformers>=4.40.0
torch>=2.1.0
tensorflow>=2.14.0
keras>=2.14.0
joblib>=1.3.2
pandas>=2.1.0
scikit-learn>=1.3.0
google-generativeai>=0.8.0
